Small edit.
This will forward all HTTP requests on the HTTPS port to the correct place.
This will fix the annoying error msg. you get when going to http://172.16.42.1:1471 after having enabled SSL.

I was a bit unsure about the variable i'm using. ($host)
The safest alternative would be $server_name, which it retrieves from the CN on the cert., but not everyone manually edits their config, and renames their pineapple.... So i landed on $host.